requirements.txt
pandas 
tqdm 
pdfplumber 
matplotlib 
torch 
transformers
pytesseract 
pdf2image 
opencv-python
numpy 
pillow
poppler-utils
PyPDF2 
pdfminer.six    